[NEW] ms reduces main() to return paths by default
This is useful in combination with the legacy front-end that adds paths
to infinite loops indicating safe termination. Such loops can be removed
for verification purposes.

This also means that ms no longer preserves program behavior. Only
program assertions (if there are any).

Old behavior is enabled with a switch. See `sea ms -h` for details