Dummy activation literal
Symbolic execution methods will take an optional activation literal.
If activation literal is given, constraints that are not definitions
will be conditioned by the activation literal.

For example, constraint C with activation literal 'a' might be
represented as 'a -> C'.

The representation is left open at the discretion of the implementation
of SymExec.