[FIX] interpret phi-nodes atomically
This is a major change in the semantics. Until now, we have assumed that
phi-nodes are processed sequentially. Thus, the evaluation of the first
phi-node affects the value of the second, etc.

However, according the LLVM semantics phi-nodes are interpreted
atomically. First, all inputs of all phi-nodes are read, and then all
outgoing values are created.

This commit changes to the new interpretation in UfoSymExec.

Note that this is only an issue when phi-nodes depend on one another. As
for example in:

  %y.0 = phi i32 [ 42, %0 ], [ %6, %3 ]
  %x.0 = phi i32 [ 84, %0 ], [ %y.0, %3 ]

The value of %x.0 is the previous value of %y.0, before the phi-node for
%y.0 is executed.

Thanks for Caterina Urban for pointing this out.