(seapp) replace truc-to-i1 with icmp-sgt
Replace the pattern

```
  x = load i8* ...
  y = trunc i8 x to i1
```

with

```
  x = load i8* ...
  y = icmp sgt i8 x, 0
```

The pattern appears when clang promotes fields of type _Bool to i8.
The trunc is added to convert i8 back to _Bool.

The two conversions are equivalent under the assumption that the field
only ever contains either 0 or 1. This is ensured by how clang generates
stores to the field.