[FIX] Bmc: using std::vector instead of SmallVector
llvm::SmallVector does not support v.push_back(v.back ()) idiom to
extend the vector using a copy constructor of the elements.

The problem is that if the push_back() makes v large enough to migrate
from stack to heap storage, v.back() is invalidated.

As a quick fix, switched to std::vector for Bmc::m_states that used this
idiom.