Adjusted scripts for new libsea-rt library
The seahorn.h is in <INSTALL_ROOT>/include/seahorn/seahorn.h
It should be included as 'seahorn/seahorn.h'

sea-clang adds <INSTALL_ROOT>/include to the default include path.

The run-time library libsea-rt is in <INSTALL_ROOT>/lib.
To generate an use counterexample harness follow the following steps

  > sea pf -m64 in.c --cex=cex.ll
  > sea clang -m64 -g -o out.bc in.c cex.ll
  > clang-mp-3.6 -m64 -o out out.bc <INSTALL_ROOT>/lib/libsea-rt.a
  > ./out
  __VERIFIER_error was executed

Option `-m64' is necessary on 64-bit platforms because the run-time
library is compiled under host architecture, but sea-clang defaults to
`-m32'.

Option `-g` is for debug information.

The last step uses system's `clang` to produce the final
executable. Ideally, it should be the same clang as used by `sea-clang`.