(dsa) prevent another source of collapsing
It is ok to collapse non array node into an array node at a non-zero
offset as long as the non-array node fits completely into the array
node.

In this case, the non-array node represents a pointer into the array
cell, but the usage so far have not converted that node into an array.

The only source of collapse is when the non-array node does not fit. In
this case, the non-array node would span multiple array cells and would
not allow to apply array-cell abstraction soundly.