(dsa) FunctionalMapper
A FunctionalMapper computes a map between two DSA sub-graphs. It assumes
that there is a functional map (each node in the source is mapped to a
unique node in the destination, but many source nodes can be mapped to
the same destination node).

This is useful to map callee sub-graph reachable from formal arguments and
globals to caller's subgraph reachable from globals and parameters.

The current implementation is very basic and likely to fail in all but
the most trivial of cases.