[FIXED] the simulation relation is built for all nodes.
During the BU pass we were keeping the map that contains the
simulation relation only for nodes that can be modified because the
global pass only uses that map to ensure a caller node can only be
mapped to one callee node. However, DSA clients (e.g.,
ShadowMemSeaDsa) needs the map for all nodes (included only read
nodes).

The solution is to add all nodes in the map during the BU pass but
during the global pass we only revisit a callsite if the relation is
not injective considering only modified nodes. Thus, we shoud not
merge more than necessary.

We also added a sanity check to make sure all nodes in the callee are
in the simulation map.