[seapp] Change default behavior of promoting assumptions
There are two kinds of assumptions: verifier.assume and llvm.assume.
They have the similar meaning, but different
clients. verifier.assume() is used by the verifier to communicate
verification assumptions. llvm.assume() is used to LLVM to communicate
invariants about the code.

We use verifier.assume() to communicate information to the verifier
that might not be true of the code (i.e., an actual assumption for
verification), or information that is true, but we want the verifier
to use it explicitly.

llvm.assume() is used to communicate true invariants between different
parts of the optimization pipeline.

Promoting verifier.assume() into llvm.assume() creates various
problems.

Sometimes, it means that the optimizer learns facts that verifier does
not want it to know.

Sometimes, the optimizer uses verifier facts unsoundly, because it
does not know that they are assumptions and not invariants.