[seapp] loop between initialization and loop conditions
Insert an empty loop while(nd()){} between the blocks where the
initialization of the loop bounds and the loop conditions happen.

The reason for this loop is to make sure that during VC encoding, the
bounds initialization and the loop conditions happen in different horn
clauses even if LBE is used.