Clean seapp. Compiles, but not tested.
Big cleanup of seapp that was long overdue.

Factored out a big if-then-else that either runs a custom pass
or the standard pipeline.

Factored out most things (like property instrumentation) out of the
default pipeline into a separate command.

Left a few things that I was not sure about in the standard pipeline.
These are commented and should be moved out.

Ensured that seapp does not include any of the headers that define
the passes that it schedules. It does not need to know their internal
structure.

Reorganized some passes in the pipeline. Nothing should be done
after inlining, unless there is a really good reason for it.
Inlining makes everything much more expensive. Most transformations
should be scheduled before inlining takes place.

None of the changes have been tested, but I believe they should not
have significant effect in most cases.