[sea]: Warn when extra arguments are ignored
Sea allows commands to accept extra undeclared arguments.
This is convenient for passing special arguments to the engine without
needing to declare each argument. However, it leads to confusion when
misspelled arguments are ignored silently.

This patch adds a warning whenever that happens so that the users are
aware which arguments from the command line were ignored.