Start of new operational semantics
This is a beginning of an attempt to rewrite operational semantics in
a cleaner way. I'm starting with BvOpSem, but if successful other
semantics will follow by refactoring and specialization.

The goal is to follow the Interpreter that's already available in LLVM
and match its semantics / behaviour as closely as possible without
sacrificing performance too much.

For now this is just very raw code, but it does compile.