Implement new OpSem API in Bv2OpSem
With the new API, the OpSemContext is passed around during VCGen.
Thus, different basic blocks know what allocations have happened
in other basic blocks.

For now, this will only work reliably in BMC mode. For CHC, the
OpSem is expected to be stateless. That is, only values() and side()
can change during an execution of a basic block.