Refactor Bv2OpSem so that Context controls sorts of registers
This is yet another refactoring step on the way of separating
general semantics from the details of the semantic machine.

Ideally, the semantic machine has full control about representation of
symbolic registers, their sorts, and mapping between LLVM types and
symbolic types. For example, should i1 map to bv(1) or to Bool.