Add cmake option SEA_ENABLE_LLD to enable lld linker
LLD linker is much faster (especially on linux) than
either bfd or gold linker. However, setting it up is
a little bit of a pain.

Add option SEA_ENABLE_LLD that is passed to LLVM as
LLVM_ENABLE_LLD. Since we use LLVM configuration for our compilation,
it means that SeaHorn will use the same settings as LLVM. Since LLMV
already provides a mechanism to use LLD, we use that.

This method is preferred to adding -fuse-ld=lld option to clang since
it is much more robust. It is easy to mis-configure clang options
into a complete mess.