Add basic WARN/ERR logging facility
RATIONALE

Currently, LOG macro is used both for persistent debug logging that is
available in release builds and to log information messages to
end-users about potential soundness issues and unsupported
constructs.

This leads to various confusion. For example, `--log=bmc` is used to
get some verbose feedback from BMC engine. However, since it is
initially intended for debugging, it also triggers an expensive
computation of unsat core that is only of interest to developers. For
example, our current BMC tests are slow because `--log=bmc` is added
and unsat core computation dominates the run time.

The WARN/ERR logging facility is to be used to report messages that
are likely to be of interest to the user. Currently, there is no
mechanism to disable the log through command line options. Support for
this is in the code, just needs to be added.

Perhaps we need to add INFO as well to log progress that is not
negative.

To use, follow the following pattern:

   WARN << "something pretty bad has happened!";

Note that there is no new line character at the end of the message.
This is intentional since WARN will add both a prefix and a suffix to
the message.

In debug builds, each warning is suffixed with a file name and line
number of where the warning was produced. This should help locating
warnings in the code and to better relate warnings to code.