Support memcpy between different memory regions
Added new mem.shadow pseudo-function shadow.mem.trsfr.load() that has exactly
the same semantics as shadow.mem.load() except that it is used to mark loads
that are part of memory transfer.

The semantics machine was extended with an additional meta-register called
memTrsfrReadReg that is loaded with a register that contains memory that is
being read by a memory transfer instruction.