Undo store arbitrary values in metadata
Arbitrary values cannot be stored as metadata.
The usage is much more constrained than I thought.

Current solution is to mark every shadow.mem instruction by
shadow.mem metadata, but not store the unique scalar inside.

Long-term, this is acceptable since the scalar is available in the
corresponding DsaNode. The clients should use DsaAnalysis for that
information in the first place.