Support registers for Functions and GlobalVariables
In llvm, Functions and GlobalVariables are considered to be constants.
They evaluate to addresses that they occupy at program start.

In our semantics, it is convenient to treat them as registers that
are initialized on program start. The work-around is to check whether
they have already been allocated registers, and if so, these registers
are used instead of the usual path of looking up constants.