Remove dependence on UnifyFunctionExitNodes from ShadowMemSeaDsa
Instead of scheduling UnifyFunctionExitNodes, assume that the returns are
unified and look for an appropriate return instruction.

LLVM Pass Manager is a big mess and relying on it is a source of many problems.
Trying to rely more on assumptions in our pipeline rather than on requesting
anything from the pass manager.