[WIP] Class hierarchy analysis for C++ programs.
The current implementation has been ported from Phasar. It is quite
limited because it relies on the fact that the constructor of a class
has not been inlined.

Currently, the class hierarchy graph is less precise that the one
produced by Phasar because it doesn't consider vtable information to
filter out some spurious edges.