Refactoring DevirtFunctions.
The code that creates bounce functions for indirect calls has not been
touched. The new thing is that DevirtFunctions is now parametric on
the method used to resolve calls (i.e., map an indirect callsite to
the set of possible callees). Currently, we only resolve by types (as
before).

This refactoring is an attempt to make easier later to plugin other
methods such as class hierarchy analysis for C++ programs or dsa.