Add new option to disable optimizer
Passing `--simplifycfg-sink-common=false` to all calls to `seapp` and `seaopt`.

When enabled, sink-common unifies instruction from multiple basic blocks into
one. This might be done at the cost of aliasing, which is causing more problems
for SeaHorn than the advantage of reducing instructions. For example,
the code of the form

```
if (x) p = v; else q = v;
```
will get optimized into

```
   w = x ? &p : &q;
  *w = v;
```

which makes `p` and `q` alias

There are comments in LLVM code that the above transformation in particular
should not happen, but it does on our examples. More investigation might
be required. For now, disabling this part since there is already a switch
to disable it that is provided by LLVM.