Lambdas for memory: --horn-bv2-lambdas=true
Use lambda expressions (instead of arrays) to model memory state.

The represenation of memory is factored into sub-classes of  OpSemMemRepr
One for arrays (old behaviour) and one for lambdas (new behaviour)

New option --horn-bv2-simplify

When enabled, expressions that are written to memory are simplified.
This can be a serious overhead for complex expressions,
especially because lambdas are beta-reduced eagerly.

Run simplifier without any limits since this works for now.
Will need to tune limit and options in the long run.
The simplifier is not caching (actually, the cache is lost
with current z3 API). Lots of things can be improved.

Replaced SpreadBytePatternAccrossWord by std::memset()

Extend mcfuzz unit tests to test labmdas