Updates comments and improves variable names
The expressions that we used to call *activation literals*
are better named *path conditions*. They don't have to be literals.
They can be arbitrary expressions. The only requirement is that a
path condition pc(bb) for a basic block bb is an expression that is
true iff and only if bb is executed.

Path conditions are used to assert that the last basic block in some
loop-free code fragment is reachable, and to create constraints that are
conditional on a basic block executed.