Enable full simplifications again
Run simplifier without any limits since this works for now.
Will need to tune limit and options in the long run.
The simplifier is not caching (actually, the cache is lost
with current z3 API). Lots of things can be improved.

Also useful debug printing to see what simplifier fails to optimize.