Factoring out ALU from OpSem
ALU is Arithmetic Logic Unit that is responsible for all arithmetic operations.
This commit starts refactoring of ALU from the main semantics.

The long term goal is to allow different variants of ALUs.

For example, we currently treat Bool differently from Bv1 by coercing
eagerly all Bv1 values to Bool. An alternative is to let the underlying
solver do this coercion if desired.

Another example is semantics based on LIA in which numbers are represented
by their arithmetic values, ignoring their target bit width.

This commit factors number of operations and introduces BvOpSemAlu that
provides bit-precise ALU. This should not change the overall behaviour.
However, the refactoring is not complete and many places in Bv2OpSem still
silently assume that values are bit-vectors.