[expr] Update conversion from Expr to z3::ast
Updated conversion to z3::ast to be less generic, but (hopefully) more efficient
and more readable.

Refactored the code to improve compilation times. Conversion code is now
compiled into an object file, so changes to it do not require recompilation of
the whole system. Similar strategies should be applied everywhere we have
excessive use of templates and excessive amount of code in header files.