[zexpr] Fix handling of some expressions
Global cache in the converter is one-to-one. For every mapping (u, v)
in the cache, both u and v are unique.

However, the relationship between Expr and z3::ast is not one-to-one.
For example, both -128:bv(8) and 128:bv(8) are mapped to 128:bv(8) in
z3::ast. This itself is correct, but causes cache to fail.

The solution is to store mappings that are not one-to-one is the local
cache. Since the local cache is one directional (from Expr to z3::ast),
this allows multiple Expr to map to the same z3::ast.