[mem-ssa] more liberal handling of memset() and memcpy()
For some reason, we felt it was important to ignore memset() and memcpy()
unless ther destination is a beginnign of a DSA node. This does not make
sense any more. The only comment is that it did not make sense in the past.

Removing and hoping that this will either work or help remember why it was
needed.