[opsem] Add COI filter to OperationalSemantics
The filter indicates which instructions are to be ignored by the semantics.

The placement of the filter is not ideal. The semantics should not care what
is to be evaluted. A beter placement for the filter would be the clients
of the semantics, like VCGen. However, current API make it inconvenient.

Might refactor at a later time as the use-case will become more clear.