[bmc] New option --horn-bmc-coi=[true|false] to enable COI
When on, applies DataFlow-based COI reduction to filter out values
that are not in dataflow COI of calls to verifier.assume().

This is only sound if all control-flow dependencies have been eliminated.
That is, there are no PHINode instructions, and no verifier.assume() statements
that are guarded by path conditions.