[opsem] optionally ignore alignment
This is a wide-byte model. Each byte address stores a word.
If all memory accesses are aligned or byte sized this is equivalent
to the usual memory model. The behavior is harder to specify clearly
when a program mixes aligned and unaligned access to the same memory object
(i.e., write a word, then read it a byte-at-a-time).

Overall, this significantly imporves scalability since each memory access in
the program is mapped to a single array access in the semantics.