[opsem] fat pointers
Memory manager that allows for fat pointers. A pointer is fat if it contains
additional bits (or slots) to store extra (meta-)data.

The currently implementation wraps pointers into structure type. The struct
currently contains a single field -- the pointer value. Hence, the pointers
are not yet really fat.

TODO:
  - add additional fields to pointers to make them trully fat
  - propagate fat during pointer manipulation (such as gep)
  - make memory regions fat, so that fat can be stored in memory