[opsem] refactor MemMgr and implment FatMem
Refactor MemMgr to remove knowledge of registers.
MemMgr should work on values. Otherwise, one mem maganger
cannot be re-used by others.

Implemented fat memory. In addition to pointers, memory is wrapped in a struct
so that each memory register can contain additional side values.

Currently, both fat-pointers and fat-memory are fat in name only.
They are wrapped by a struct, but the struct contains only one field -- the
pointer or the memory value.

The main issue was in re-factoring the code to remove assumption on pointer type
and memory value type. Extending pointers and memory with actual fat should be
fairly straightforward.