FIX: blocking clause generation in path-bmc engine.
Until now Clam called directly sea-dsa and leveraged it to translate
LLVM memory instructions to Crab arrays. This means that sea-dsa
analysis was being called twice when SeaHorn was running with Clam:
one by ShadowMem and the other by Clam. Although potentially
inefficient, this worked when SeaHorn loads Clam invariants but it was
not good enough for the path-bmc engine.

This bmc engine uses Clam/Crab to check whether a path is satisfiable
or not. If unsat, then it produces a blocking clause. Generating a
blocking clause is tricky because the SMT encoding of the BMC problem
and the Crab CFG generated by Clam must be very close since the BMC
engine must map back the blocking clause generated by Crab into the
SMT encoding. It turned out that the blocking clause generation was
unsound because Crab did not preserve memory SSA form while the SMT
encoding does. In particular, the encoding of Crab array
write/selects.

The solution was to have an alternative Clam encoding to Crab CFGs
that preserves the memory SSA form generated by the ShadowMem pass
from sea-dsa. That is, now both Clam and SeaHorn uses ShadowMem to
encode memory SSA form into Crab CFG and SMT encoding, respectively.

The horn clause-based engine still uses Clam calling directly sea-dsa.
This is OK but we want to switch soon to ShadowMem. The current
bottleneck is that the Clam translation using ShadowMem is only
implemented for the intra-procedural case.