Moved caching of bounce functions into the particular resolver.
This is in preparation for supporting sea-dsa as a resolver for
indirect calls.  The code was assuming that a bounce function can be
reused if its type signature matches the callsite. With sea-dsa we
want to reuse a bounce function if they have the same type signature
and they cover the same set of targets.