[NEW] VCGen encoding that exploits finite memory used by functions.
The new VCGen encoding is disabled by default. Add the option
--horn-inter-proc-mem to enable it.

This is a new VCGen encoding that leverages sea-dsa to identify per
callsite whether the callee uses a finite amount of memory. If this is
the case, the caller only passes that finite memory to the
callee. This can have significant impact on the back-end solver when
generating summaries.

Current VCGen encoding always passes (at best) one unbounded SMT array
per argument to the callee. That array, let us call it ComplexArr, can
represent a large, complex datastructure. However, the callee might
access to a small part of it.  As a result, the callee summary needs
to express that the callee modifies a small part of ComplexArr and
leaves unmodified the rest. This kind of summaries are hard to
generate for the back-end solvers because typically involve
quantifiers.
 
With the new VCGen if the callee uses a bounded amount of memory, the
encoding passes a _different_ array to the callee containing only the
memory regions accessed by the callee. This eliminates the need in the
summaries to say which part of the complex datastructure remains
intact since ComplexArr is not passed anymore the callee. This makes
much simpler the summaries generated by the back-end solver.