[test-opsem] Adapt nham.1.ll test
Something (potentially in rewriter) has changed in Z3 that makes the
original test stuck in initial pre-processing. This requires further
investigation since pre-processing is assumed to be inexpensive.

Was not an issue for z3/deep_space

Checked with current release z3/4.8.7 and current master z3/4.8.8.
Same problem with both.

Applying pre-processing earlier during construction of the VC helps.