fix(z3): Support only single query with arity 0
The interface of Z3 fixedpoint module has changed. The old way to support
multiple queries and queries with non-zero arity no longer works.
Since this is not used by SeaHorn, current fix is to disable the old
code that no longer works as intended.

This restricts functionality, but hopefully in a way that does not affect any
existing client.