feat(analysis): New pass for accessing sea builtin functions
Problem: we assume implicit presence of SeaHorn specific builtin functions such
as verifier.error() and verifier.assume(). However, there is no centralized
place for creating and accessing them.

Solution: A data structure SeaBuiltinsInfo and an immutable pass
SeaBuiltinsInfoWrapperPass to access it in passes.

TODO: Convert the code to use SeaBuiltinsInfo (or SBI) instead of creating or
assuming existence of seahorn-specific functions