test(devirt): adapt test
The new code that promotes indirect calls to direct calls does not add
any bounce functions. This makes the outcome of devirt_cha_02.false.cpp to
change from sat to unsat.

The cex depends on the fact that a function is recursive but LLVM
optimizations can tell now that the recursive call does not terminate
(due to lack of bounce functions) and it get rids of it.