Use llvm10 utilities to promote indirect calls (#300)
* refactor(devirt): use llvm 10 utilities to promote indirect calls.

llvm 10 provides utilities to replace indirect calls with direct
calls. We use now these utilities to replace the old code that adds
bounce (trampoline) functions to make easier code maintenance.

We keep the code that generates bounce functions for debugging.  There
is a pragma USE_BOUNCE_FUNCTIONS in DevirtFunctions.h that can enable
old behavior.

* test(devirt): adapt test

The new code that promotes indirect calls to direct calls does not add
any bounce functions. This makes the outcome of devirt_cha_02.false.cpp to
change from sat to unsat.

The cex depends on the fact that a function is recursive but LLVM
optimizations can tell now that the recursive call does not terminate
(due to lack of bounce functions) and it get rids of it.

* style(devirt): clang-format

* refactor(cmake): USE_BOUNCE_FUNCTIONS as cmake option