fix: handling of getCalledFunction()
in LLVM, getCalledFunction returns a nullptr if the function is called through a
cast. We like, in this case, for the cast to be stripped to separate this from
indirect calls.

in LLVM, llvm::AbstractCallSite does the same thing, but we don't need the other
aspects of it right now.