feat(bvopsem2): mixin for OpSemMemManager
It is convenient to write MemManagers against a well defined type.
For example, if a pointer is a record, then it should be represented
by a struct, not by an Expr.

However, the clients (i.e., OperationalSemantics) expect everything to be an
Expr. The mixin adapts a memory manager with custom types to what is expected
by the clients.