feat(hornify): bit-precise CHC
New option `sea pf --bv-chc` produces bit-precise CHC.

Bit-precise semantics is based on (old) BvOpSem. This is the same
semantics that is used by default by the BMC engine and for bit-precise
validation of counterexamples.

Note that while the semantics is bit-precise, memory modelling is not
precise (this is being worked out in bvopsem2). Thus, the user is expected to
provide models of the allocator functions, or ensure that pointer analysis is
powerful enough to disambiguate sufficiently all allocation sites.