fix(klee): clear metadata before generating a body for a function
Function declarations and function definitions have different metadata in LLVM.
By not clearing the metadata, the metadata of the declaration is inherited
by the definition. This causes clang++ to crash while linking such code.

The underlying bug is a null pointer deref in clang (which we can report).
This is in part a work around, in part, following the implicit contract on
metadata.