refactor(DevirtFunction): use of --devirt-functions-allow-incomplete flag.
By default, an indirect call is replaced with direct calls even if
sea-dsa might not know all possible callees. This is potentially
unsound because of the missing unknown callees.

If --devirt-functions-allow-incomplete is enabled then
devirtualization will add a direct call to all possible callees plus
one extra call to the original callsite.

Note that --devirt-functions=types already worked in this way.