feat(seahorn): enable symbolize bounds of constant loops
We had this pass for a while but it was not wired by default.
The option to start it no longer works after we switched to LLVM10

I've added it into the pipeline. It is always enabled in CHC mode.
I do not see any good reason to not enable it. It will apply only to loops that
have constant trip count after all of the optimizations have been applied.
So it does not prevent bounded unrolling that might be useful on its own.