feat(opsem): memcpy with symbolic length
Fully supported with lambdas and raw memory manager.
Not supported with arrays.
Probably not yet supported with wide memory managers.

Probably buggy with word_size > 1 and length that is not multiple of word_size.
We assume that src and dst pointers are aligned, but if length is not a multiple
of word_size then the last word must be reconstructed from the bytes copied from
the source and bytes in the destination. Not sure we want to deal with this or
require word_size of 1 for examples that require such reasoning.