chore(seapp): hack to support bcmp
By default, all functions that are not used are removed from the module.
This is a problem for functions to which calls are introduced by the optimizer.

In particular, memcmp is optimized into bcmp if possible. So a call to memcmp
is removed and a call to bcmp is added.

This hack treats bcmp as special and does not mark it as internal.
Better long term solution is needed.