feat(opsem): support memmove()
Our semantics of memcpy are essentially memove. The implementation forwards
memmove to memcpy. Although, it really should be the other way around.

memmove() allows source and destination to overlap. In our case, we have a
snapshot of the source and copy that into dest. Thus, overlap is not a problem.

memcpy() says that in case input pointers overlap, its behaviour is undefined.
So it is safe to implement memcpy() using memmove(). This is not done in
practice because memmove() tends to be more expensive.

We can add a property to check that memcpy() applications satisfy its pre-conditions.