feat(sea): do not rewrite away arithmetic overflow intrinsics
The pp pipeline used to always rewrite arithmetic overflow intrinsics with their
operators and assume that overflow never happens. This was needed for CHC engine
and clam that operate on infinite precision numbers. This is a problem for
bvopsem that supports those operators.

Made this optional in seapp and added a corresponding option to `sea pp`