fix(opsem): memset(0) in ExtraWideMemMgr
MemSet was applied only on the raw part of memory, ignoring the offset.
This is based on the assumption that memset will not create any valid pointers,
thus we can leave offset alone in some undefined state.

There is one exception though. `memset(0)` is a common idiom to initialize
memory, including setting all pointers to `NULL`. I've added special handling
for this case.