feat(opsem): handle very simple inline asm
These are instructions we will handle
1) %_57 = call i64 asm sideeffect "", "=r,0,~{dirflag},~{fpsr},~{flags}"(i64 %_14) #7, !dbg !504, !srcloc !505 @ _55 in main
2) call void asm sideeffect "", "r,~{memory},~{dirflag},~{fpsr},~{flags}"(i8* %7) #5, !dbg !102, !srcloc !103

Background:
There are three kinds of situations with inline assembly in aws-c-common
1. Code to trick the compiler to not optimize. Above two examples fall in this category
2. Code to do things efficiently :e.g. add-with_saturation, aws_hton64 . in CBMC they delegate to a C function instead of assembly - i believe we should do/already be doing the same
3. Code to get system info using special registers e.g. CPU-ID on intel processors. I don't think we should be formally verifying this kind of "system code".

Regarding the ~{...} registers... i believe that LLVM does not want to use those registers for storing input/output values and so it makes a note for itself. We will ignore this information.
Regarding the "sideeffect" tag, I believe LLVM thinks the ~{dirflag},~{fpsr},~{flags} must always be changed and so notes that this function has side effects. We will ignore this information.