fix(transforms): do not externalize some functions
A common pattern is to cast a function before calling it.
Do not externalize such functions.

In general, we really should only externalize functions whose address
escapes (i.e., stored into memory or passed as an argument to unknown function).
As a proxy, we externalize all functions whose address is taken.

This fix throttles the approximation a little bit.