fix(mixsm): segfault on a pathological example
This is a work-around. Full fix might still be required.

Mixed Semantics reduces the program to basic blocks that are on a path from
entry to error blocks.

If an error block is unreachable, it might be removed by
removeUnreachableBlocks() and will become invalid during the marking process.

A proper fix is to ensure that all blocks that are removed by
removeUnreachableBlocks() are also removed from the list of error blocks. This
requires a rewrite of removeUnreachableBlocks() and more testing.

Currently, the error is only triggered on mcfuzz test cases with undefined
behaviour (thus many unreachable blocks). The current work around is to not mark
any error blocks that do not have predecessors.