fix(opsem): treat store of double/float to memory as noop
We do not support float/double types at the moment. The previous behaviour was
too conservative -- a write of double/float was treated as havoc of the memory
region to which the write is done.

The current change warns and treats such a write as a noop. This is sound as
long as
  - reads of double/float are also treated as noop
  - writes of double/float do not override integer/pointers

The last point is not allowed under strict aliasing semantics of C anyhow.