feat(seapp): option to disable promotion of undef to nondet
Option --promote-nondet-undef of seapp controls whether `undef` in bitcode is
promoted to non-determinism. Currently, the initial value is true.

sea script has to be extended to allow users to control setting the option to
false.

The option is useful to validate that any resolution of `undef` is a well
defined program. However, it is inconvenient when verification is done for some
other property, under assumption that `undef` does not cause any undefined
behaviour.