feat(seahorn): generated rules for synthesis (small step)
When a synthesis intrinsic appears in a block, seahorn replaces the intrinsic
with the wrapped call. When sea.synth.assert appears in a block, a second rule
is generated in which the block is truncated before the intrinsic, and then the
wrapepd call appears on the right hand side. For a block (1), the rules (2) and
(3) are generated.

1. pre && (tr1 && assert(fn) && tr2) => post
2. pre && tr1 => fn
3. pre && tr1 && fn && tr2 => post

Note: getPartialFnsToSynth and expandEdgeFilter are factored for large step.