feat(seahorn): vcgen can omit dst (phi nodes) on cut point edges
VCGen takes as input a CpEdge, assembles intructions for each bb along the edge,
and then assembles phi nodes for the destination of the edge. However, large
step synthesis requires that VCGen support edges whose destination overlaps with
a bb along the edge. In this case, all phi nodes should be omitted.

The solution is to add a flag to genVcForCpEdge that enables Phi node asesmbly.
The flag is raised by default, to retain existing behaviour.