feat(sea-py): processing of extra options in yama command
Extra options in yama command override options in the yaml files.

Previously, extra options where simply added to the command line. This causes an
error if an option is already specified in a yaml file and is not allowed to
appear multiple times.

The new behaviour merges options with the yaml options by overriding them. This
is the natural expected behaviour of additional command line options.