fix: prototype of __VERIFIER_assume
The function `__VERIFIER_assume` is defined in SV-COMP as

```
void __VERIFIER_assume(int)
```

We prefer it to take `bool` instead of int. However, it is not legal to redefine it
locally since this contradicts with what other tools might expect.

This commit introduces our own function

```
void __SEA_assume(bool)
```

and redefines `assume` to resolve to it.

There are no expected change to any clients.