fix(path-bmc): crab blocking path generation.
Conceptually, these are the steps we need to do when Crab proves that
a cex is spurious:

(1) Produce a LLVM cex from the set of Crab statements that proves that
the cex is spurious.

(2) Produce encoding of the LLVM cex using VCGen and extract from
there the boolean literals (i.e., boolean abstraction of the path).

(3) Refine the boolean abstraction by eliminating the boolean
abstraction of the path produced by Step 2.

The implementation skips step 2 and goes directly from the set of Crab
statements to the refinement of the boolean abstraction, but for discussion,
let us assume we follow steps 1-3.

Both Crab and VCGen uses the same memory partitions provided by
sea-dsa, but VCGen uses multiple versions of the same partition (a new
version each time a partition is redefined) while Crab uses one single
version per partition. This means that VCGen is based on memory
SSA-form while CrabIR is not. This caused problems in the past.

This fix addresses the case where a memory read is part of the set of
Crab statements in step 1. In that case, we also need to include in
the LLVM cex all the PHI nodes that connect that memory read with the
last memory write on the same partition (i.e., the whole memory
use-def chain). This is done by encodeUseDefChain. The rest of changes
are related to code factorization.