fix(path-bmc): do not use LiveSymbols.
LiveSymbols analysis creates its own symbolic execution contexts which
are not in sync with the ones computed by the BMC engine (via its
opsem or opsem2). Moreover, we do not really need it. Live symbols are
used to prune Crab invariants and Crab already computes live symbols
so it makes more sense to ask directly Crab.