feat: print z3 solver using native printer
SeaHorn has several custom printers to print a state of a ZSolver object.
The reasons are mostly legacy, since the printing of individual expressions is
delegated to Z3 in all cases. The difference is only in how declarations are
produced and whether assertions are merged together.

This commit adds a compile-time option ZSOLVER_NATIVE_PRINT that, when set to
true, enables native z3 printer that prints solver object as an SMT-LIB
benchmark, including all the necessary declarations.