refactor(opsem2): new crab option for is_deref
The new option is horn-bv2-crab-check-is-deref.  We have now two
options horn-bv2-crab-lower-is-deref and
horn-bv2-crab-check-is-deref. Both options use Crab to prove is_deref
intrinsics so that the BMC engine does not need to do it. The former
adds an instrumentation at the CrabIR level to reason about is_deref
while the latter does not modify the CrabIR and then it is the
abstract domain which needs to understand the intrinsics.