fix(smc): disable sat SMC tests that are now unsat
With change in alignment, the SMC tests no longer fail since there is
sufficiently many bytes allocated after a field.

Since we are not maintaining SMC anymore (at least not in the near future),
I'm disabling the tests rather than trying to rework them to cause the bug to
appear even with the change in alignment.