feat(py): add a new sea option for crab in seapp
If we disable crab feature in seapp, we should either:
(1) run seapp as no-op; (2) remove the feature from the pipeline.
However, for (1), this will cause seapp to run default pre-processing pipeline.
For (2), instead of changing original pipeline, we add a new sea option 'fpcf'
to run crab lower is_deref in preprocessing.